+++ src/main/java/racingcar/domain/RacingCarGame.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 import racingcar.exception.RacingCarException;
